+++ check_raid.pl
@@ -714,6 +714,7 @@ sub parse {
 
 		push(@md, { %md } ) if %md;
 	}
+	close $fh;
 
 	# One of the arrays is in checking state, which could be because there is a scheduled sync of all MD arrays
 	# In such a case, all of the arrays are scheduled to by checked, but only one of them is actually running the check
@@ -727,7 +728,6 @@ sub parse {
 			}
 		}
 	}
-	close $fh;
 
 	return wantarray ? @md : \@md;
 }
@@ -987,18 +987,26 @@ sub parse_bbu {
 
 	return undef unless $this->bbu_monitoring;
 
+	my %default_bbu = (
+		name => undef, state => '???', charging_status => '???', missing => undef,
+		learn_requested => undef, replacement_required => undef,
+		learn_cycle_requested => undef, learn_cycle_active => '???',
+		pack_will_fail => undef, temperature => undef, temperature_state => undef,
+		voltage => undef, voltage_state => undef
+	);
+
 	my (@bbu, %bbu);
 	my $fh = $this->cmd('battery');
 	while (<$fh>) {
+		# handle when bbu status get gives an error. see issue #32
+		if (my($s) = /Get BBU Status Failed/) {
+			last;
+		}
+
 		if (my($s) = /BBU status for Adapter: (.+)/) {
 			push(@bbu, { %bbu }) if %bbu;
-			%bbu = (
-				name => $s, state => '???', charging_status => undef, missing => undef,
-				learn_requested => undef, replacement_required => undef,
-				learn_cycle_requested => undef, learn_cycle_active => undef,
-				pack_will_fail => undef, temperature => undef, temperature_state => undef,
-				voltage => undef, voltage_state => undef
-			);
+			%bbu = %default_bbu;
+			$bbu{name} = $s;
 			next;
 		}
 #=cut
@@ -1053,10 +1061,11 @@ sub parse_bbu {
 			$bbu{voltage_state} = $s;
 			next;
 		}
-	}
-	push(@bbu, { %bbu }) if %bbu;
 
+	}
 	$this->critical unless close $fh;
+
+	push(@bbu, { %bbu }) if %bbu;
 
 	return \@bbu;
 }
@@ -1418,8 +1427,8 @@ sub check {
 
 	my $read = $this->cmd('container list', \$write);
 	while (<$read>) {
- 		# 0    Mirror  465GB            Valid   0:00:0 64.0KB: 465GB Normal                        0  032511 17:55:06
- 		# /dev/sda             root             0:01:0 64.0KB: 465GB Normal                        1  032511 17:55:06
+		# 0    Mirror  465GB            Valid   0:00:0 64.0KB: 465GB Normal                        0  032511 17:55:06
+		# /dev/sda             root             0:01:0 64.0KB: 465GB Normal                        1  032511 17:55:06
 		if (my($dsk, $stat) = /(\d:\d\d?:\d+)\s+\S+:\s?\S+\s+(\S+)/) {
 			next unless $this->valid($dsk);
 			$dsk =~ s#:#/#g;
@@ -1457,7 +1466,7 @@ sub program_names {
 
 sub commands {
 	{
-		'status' => ['-|', '@CMD', '-i $id'],
+		'status' => ['-|', '@CMD', '-i', '$id'],
 		'get_controller_no' => ['-|', '@CMD', '-p'],
 		'sync status' => ['-|', '@CMD', '-n'],
 	}
@@ -1549,7 +1558,7 @@ sub parse {
 				flags => [ split ' ', $flags ],
 			};
 		} else {
-			warn "mpt unparsed: [$_]\n";
+			warn "mpt unparsed: [$_]";
 			$this->unknown;
 		}
 	}
@@ -1566,10 +1575,10 @@ sub parse {
 			}
 			# mpt-status.c GetResyncPercentage
 			# scsi_id:0 70%
-			elsif (my($scsi_id, $percent) = /^scsi_id:(\d)+ (\d+)%/) {
+			elsif (my($scsi_id, $percent) = /^scsi_id:(\d+) (\d+)%/) {
 				$pd{$scsi_id}{resync} = int($percent);
 			} else {
-				warn "mpt: [$_]\n";
+				warn "mpt unparsed: [$_]";
 				$this->unknown;
 			}
 		}
@@ -1769,7 +1778,7 @@ sub parse {
 
 					$pd{$p{id}} = { %p };
 				} else {
-					warn "[$section] [$_]\n";
+					warn "[$section] [$_]";
 					$this->unknown;
 				}
 
@@ -1791,7 +1800,7 @@ sub parse {
 				} elsif (/^$/) {
 					$ld{$l{number}} = { %l };
 				} else {
-					warn "[$section] [$_]\n";
+					warn "[$section] [$_]";
 					$this->unknown;
 				}
 
@@ -1809,7 +1818,7 @@ sub parse {
 						$ad{$a{number}} = { %a };
 					}
 				} else {
-					warn "[$section] [$_]\n";
+					warn "[$section] [$_]";
 					$this->unknown;
 				}
 
@@ -2033,7 +2042,7 @@ sub check {
 			if ($s eq 'OK') {
 				push(@cstatus, "$u:$s");
 
-			} elsif ($s =~ 'INITIALIZING|MIGRATING') {
+			} elsif ($s =~ /INITIALIZING|MIGRATING/) {
 				$this->warning;
 				push(@cstatus, "$u:$s $p2");
 
@@ -2293,14 +2302,14 @@ sub parse_config {
 					$c{battery_time_full} = "${d}d${h}h${m}m";
 
 				} else {
-					warn "Battery not parsed: [$_]\n";
+					warn "Battery not parsed: [$_]";
 				}
 
 			} elsif ($subsection eq 'Controller ZMM Information') {
 				if (my($bs) = /^\s+Status\s*:\s*(.*)$/) {
 					$c{zmm_status} = $bs;
 				} else {
-					warn "ZMM not parsed: [$_]\n";
+					warn "ZMM not parsed: [$_]";
 				}
 
 			} elsif ($subsection eq 'Controller Version Information') {
@@ -2315,7 +2324,7 @@ sub parse_config {
 			} elsif ($subsection eq 'Controller Vital Product Data') {
 				# not parsed yet
 			} else {
-				warn "SUBSECTION of [$section] NOT PARSED: [$subsection] [$_]\n";
+				warn "SUBSECTION of [$section] NOT PARSED: [$subsection] [$_]";
 			}
 
 		} elsif ($section eq 'Physical Device information') {
@@ -2332,7 +2341,7 @@ sub parse_config {
 				} elsif (/No physical drives attached/) {
 					# ignored
 				} else {
-					warn "Unparsed Physical Device data: [$_]\n";
+					warn "Unparsed Physical Device data: [$_]";
 				}
 			} else {
 				if (my($ps) = /Power State\s+:\s+(.+)/) {
@@ -2398,8 +2407,10 @@ sub parse_config {
 					# not parsed yet
 				} elsif (/MaxCache (Capable|Assigned)\s+:\s+(.+)/) {
 					# not parsed yet
+				} elsif (/Power supply \d+ status/) {
+					# not parsed yet
 				} else {
-					warn "Unparsed Physical Device data: [$_]\n";
+					warn "Unparsed Physical Device data: [$_]";
 				}
 			}
 
@@ -2436,7 +2447,7 @@ sub parse_config {
 		} elsif ($section =~ /MaxCache 3\.0 information/) {
 			# not parsed yet
 		} else {
-			warn "NOT PARSED: [$section] [$_]\n";
+			warn "NOT PARSED: [$section] [$_]";
 		}
 	}
 	close $fh;
@@ -2550,6 +2561,7 @@ sub check {
 
 	# check for physical devices
 	my %pd;
+	my $pd_resync = 0;
 	for my $ch (@{$data->{physical}}) {
 		for my $pd (@{$ch}) {
 			# skip not disks
@@ -2558,12 +2570,13 @@ sub check {
 
 			if ($pd->{status} eq 'Rebuilding') {
 				$this->resync;
+				$pd_resync++;
 
 			} elsif ($pd->{status} eq 'Dedicated Hot-Spare') {
 				$this->spare;
 				$pd->{status} = "$pd->{status} for $pd->{spare}";
 
-			} elsif ($pd->{status} !~ '^Online') {
+			} elsif ($pd->{status} !~ '^Online|Hot[- ]Spare') {
 				$this->critical;
 			}
 
@@ -2576,7 +2589,11 @@ sub check {
 	for my $ld (@{$data->{logical}}) {
 		next unless $ld; # FIXME: fix that script assumes controllers start from '0'
 
-		$this->critical if $ld->{status} !~ /Optimal|Okay/;
+		if ($ld->{status} eq 'Degraded' && $pd_resync) {
+			$this->warning;
+		} elsif ($ld->{status} !~ /Optimal|Okay/) {
+			$this->critical;
+		}
 
 		my $id = $ld->{id};
 		if ($ld->{name}) {
@@ -3138,6 +3155,7 @@ sub commands {
 	{
 		'controller list' => ['-|', '@CMD', 'LIST'],
 		'controller status' => ['-|', '@CMD', '$controller', 'STATUS'],
+		'device status' => ['-|', '@CMD', '$controller', 'DISPLAY'],
 	}
 }
 
@@ -3150,6 +3168,7 @@ sub sudo {
 	(
 		"CHECK_RAID ALL=(root) NOPASSWD: $cmd LIST",
 		"CHECK_RAID ALL=(root) NOPASSWD: $cmd * STATUS",
+		"CHECK_RAID ALL=(root) NOPASSWD: $cmd * DISPLAY",
 	);
 }
 
@@ -3161,9 +3180,12 @@ sub detect {
 	my $fh = $this->cmd('controller list');
 
 	my $success = 0;
+	my $state="";
+	my $noctrlstate="No Controllers";
 	while (<$fh>) {
 		chomp;
-		#		  Adapter     Vendor  Device                        SubSys  SubSys
+
+		#         Adapter     Vendor  Device                        SubSys  SubSys
 		# Index    Type          ID      ID    Pci Address          Ven ID  Dev ID
 		# -----  ------------  ------  ------  -----------------    ------  ------
 		#   0     SAS2008     1000h    72h     00h:03h:00h:00h      1028h   1f1eh
@@ -3171,10 +3193,30 @@ sub detect {
 			push(@ctrls, $c);
 		}
 		$success = 1 if /SAS2IRCU: Utility Completed Successfully/;
+
+		# handle the case where there's no hardware present.
+		# when there is no controller, we get
+		# root@i41:/tmp$ /usr/sbin/sas2ircudsr LIST
+		# LSI Corporation SAS2 IR Configuration Utility.
+		# Version 18.00.00.00 (2013.11.18)
+		# Copyright (c) 2009-2013 LSI Corporation. All rights reserved.
+
+		# SAS2IRCU: MPTLib2 Error 1
+		# root@i41:/tmp$ echo $?
+		# 1
+
+		if ( /SAS2IRCU: MPTLib2 Error 1/ ) {
+			$state = $noctrlstate;
+			$success = 1 ;
+		}
+
 	}
 
 	unless (close $fh) {
-		$this->critical;
+		#sas2ircu exits 1 (but close exits 256) when we close fh if we have no controller, so handle that, too
+		if ( $? != 256 && $state eq $noctrlstate ) {
+			$this->critical;
+		}
 	}
 	unless ($success) {
 		$this->critical;
@@ -3183,16 +3225,22 @@ sub detect {
 	return wantarray ? @ctrls : \@ctrls;
 }
 
+sub  trim { my $s = shift; $s =~ s/^\s+|\s+$//g; return $s };
+sub ltrim { my $s = shift; $s =~ s/^\s+//;       return $s };
+sub rtrim { my $s = shift; $s =~ s/\s+$//;       return $s };
+
 sub check {
 	my $this = shift;
 
 	my @ctrls = $this->detect;
 
 	my @status;
+	my $numvols=0;
 	# determine the RAID states of each controller
 	foreach my $c (@ctrls) {
 		my $fh = $this->cmd('controller status', { '$controller' => $c });
 
+		my $novolsstate="No Volumes";
 		my $state;
 		my $success = 0;
 		while (<$fh>) {
@@ -3201,12 +3249,129 @@ sub check {
 			# match adapter lines
 			if (my($s) = /^\s*Volume state\s*:\s*(\w+)\s*$/) {
 				$state = $s;
+				$numvols++;
 				if ($state ne "Optimal") {
 					$this->critical;
 				}
 			}
 			$success = 1 if /SAS2IRCU: Utility Completed Successfully/;
+
+			##handle the case where there are no volumes configured
+			#
+			# SAS2IRCU: there are no IR volumes on the controller!
+			# SAS2IRCU: Error executing command STATUS.
+			#
+
+			if ( /SAS2IRCU: there are no IR volumes on the controller!/ ) {
+				#even though this isn't the last line, go ahead and set success.
+				$success = 1;
+				$state = $novolsstate;
+			}
+
 		}
+
+		unless (close $fh) {
+			#sas2ircu exits 256 when we close fh if we have no volumes, so handle that, too
+			if ( $? != 256 && $state eq $novolsstate ) {
+				$this->critical;
+				$state = $!;
+			}
+		}
+
+		unless ($success) {
+			$this->critical;
+			$state = "SAS2IRCU Unknown exit";
+		}
+
+		unless ($state) {
+			$state = "Unknown Error";
+		}
+
+		my $finalvolstate=$state;
+		#push(@status, "ctrl #$c: $numvols Vols: $state");
+
+
+		#####  now look at the devices.
+		# Device is a Hard disk
+		#   Enclosure #                             : 2
+		#   Slot #                                  : 0
+		#   SAS Address                             : 500065b-3-6789-abe0
+		#   State                                   : Ready (RDY)
+		#   Size (in MB)/(in sectors)               : 3815447/7814037167
+		#   Manufacturer                            : ATA
+		#   Model Number                            : ST4000DM000-1F21
+		#   Firmware Revision                       : CC52
+		#   Serial No                               : S30086G4
+		#   GUID                                    : 5000c5006d27b344
+		#   Protocol                                : SATA
+		#   Drive Type                              : SATA_HDD
+
+		$fh = $this->cmd('device status', { '$controller' => $c });
+		$state="";
+		$success = 0;
+		my $enc="";
+		my $slot="";
+		my @data;
+		my $device="";
+		my $numslots=0;
+		my $finalstate;
+		my $finalerrors="";
+
+		while ( my $line = <$fh> ) {
+			chomp $line;
+			# Device is a Hard disk
+			# Device is a Hard disk
+			# Device is a Enclosure services device
+			#
+			#lets make sure we're only checking disks.  we dont support other devices right now
+			if ( "$line" eq 'Device is a Hard disk' ) {
+				$device='disk';
+			} elsif ( $line =~ /^Device/ )  {
+				$device='other';
+			}
+
+			if ( "$device" eq 'disk' ) {
+				if ( $line =~ /Enclosure #|Slot #|State / ) {
+					#find our enclosure #
+					if ( $line =~ /^  Enclosure # / ) {
+						@data = split /:/, $line;
+						$enc=trim($data[1]);
+						#every time we hit a new enclosure line, reset our state and slot
+						undef $state;
+						undef $slot;
+					}
+					#find our slot #
+					if ( $line =~ /^  Slot # / ) {
+						@data = split /:/, $line;
+						$slot=trim($data[1]);
+						$numslots++
+					}
+					#find our state
+					if ( $line =~ /^  State / ) {
+						@data = split /:/, $line;
+						$state=ltrim($data[1]);
+
+						#for test
+						#if ($numslots == 10 ) { $state='FREDFISH';}
+
+						#when we get a state, test on it and report it..
+						if ( $state =~ /Optimal|Ready/ ) {
+							#do nothing at the moment.
+						} else {
+							$this->critical;
+							$finalstate=$state;
+							$finalerrors="$finalerrors ERROR:Ctrl$c:Enc$enc:Slot$slot:$state";
+						}
+					}
+				}
+			}
+
+			if ( $line =~ /SAS2IRCU: Utility Completed Successfully/) {
+				$success = 1;
+			}
+
+		} #end while
+
 
 		unless (close $fh) {
 			$this->critical;
@@ -3222,7 +3387,19 @@ sub check {
 			$state = "Unknown Error";
 		}
 
-		push(@status, "ctrl #$c: $state")
+		unless($finalstate) {
+			$finalstate=$state;
+		}
+
+		#per controller overall report
+		#push(@status, ":$numslots Drives:$finalstate:$finalerrors");
+		push(@status, "ctrl #$c: $numvols Vols: $finalvolstate: $numslots Drives: $finalstate:$finalerrors:");
+
+	}
+
+	##if we didn't get a status out of the controllers and an empty ctrls array, we must not have any.
+	unless (@status && @ctrls) {
+		push(@status, "No Controllers");
 	}
 
 	return unless @status;
@@ -3372,9 +3549,9 @@ sub check {
 		return;
 	}
 
-
 	# Scan logical drives
-	while (my($target, $model) = each %targets) {
+	for my $target (sort {$a cmp $b} keys %targets) {
+		my $model = $targets{$target};
 		# check each controllers
 		my $fh = $this->cmd('logicaldrive status', { '$target' => $target });
 
@@ -3589,12 +3766,16 @@ sub program_names {
 
 sub commands {
 	{
-		'read' => ['-|', '@CMD', '-r'],
+		'dmraid' => ['-|', '@CMD', '-r'],
 	}
 }
 
 sub active ($) {
 	my ($this) = @_;
+
+	# easy way out. no executable
+	return 0 unless -e $this->{commands}{dmraid}[1];
+
 	# check if dmraid is empty
 	return keys %{$this->parse} > 0;
 }
@@ -3613,7 +3794,7 @@ sub parse {
 	my $this = shift;
 
 	my (%arrays);
-	my $fh = $this->cmd('read');
+	my $fh = $this->cmd('dmraid');
 	while (<$fh>) {
 		chomp;
 		next unless (my($device, $format, $name, $type, $status, $sectors) = m{^
@@ -3687,9 +3868,9 @@ use Getopt::Long;
 
 my ($opt_V, $opt_d, $opt_h, $opt_W, $opt_S, $opt_p, $opt_l);
 my (%ERRORS) = (OK => 0, WARNING => 1, CRITICAL => 2, UNKNOWN => 3);
-my ($VERSION) = "3.0.6";
+my ($VERSION) = "3.1.1";
 my ($message, $status, $perfdata, $longoutput);
-my ($opt_O) = $ERRORS{UNKNOWN};
+my ($noraid_state) = $ERRORS{UNKNOWN};
 
 #####################################################################
 $ENV{'BASH_ENV'} = '';
@@ -3705,7 +3886,7 @@ sub find_file {
 
 sub print_usage() {
 	print join "\n",
-	"Usage: check_raid [-h] [-V] [-S] [-O] [list of devices to ignore]",
+	"Usage: check_raid [-h] [-V] [-S] [list of devices to ignore]",
 	"",
 	"Options:",
 	" -h, --help",
@@ -3744,6 +3925,56 @@ https://github.com/glensc/nagios-plugin-check_raid
 	print_usage();
 }
 
+# return first "#includedir" directive from $sudoers file
+sub parse_sudoers_includedir {
+	my ($sudoers) = @_;
+
+	open my $fh, '<', $sudoers or die "Can't open: $sudoers: $!";
+	while (<$fh>) {
+		if (my ($dir) = /^#includedir\s+(.+)$/) {
+			return $dir;
+		}
+	}
+	close $fh or die $!;
+
+	return undef;
+}
+
+# return size of file
+# does not check for errors
+sub filesize {
+	my ($file) = @_;
+	return (stat($file))[7];
+}
+
+# get contents of a file
+sub cat {
+	my ($file) = @_;
+	open(my $fh, '<', $file) or die "Can't open $file: $!";
+	local $/ = undef;
+	local $_ = <$fh>;
+	close($fh) or die $!;
+
+	return $_;
+}
+
+# return TRUE if files are identical
+# returns FALSE if any of the files is missing
+sub filediff {
+	my ($file1, $file2) = @_;
+
+	return unless -f $file1;
+	return unless -f $file2;
+
+	return unless filesize($file1) != filesize($file2);
+
+	return cat($file1) eq cat($file2);
+}
+
+# update sudoers file
+#
+# if sudoers config has "#includedir" directive, add file to that dir
+# otherwise update main sudoers file
 sub sudoers {
 	my ($dry_run) = @_;
 
@@ -3795,6 +4026,15 @@ sub sudoers {
 	die "Unable to write to sudoers file '$sudoers'.\n" unless -w $sudoers;
 	die "visudo program not found\n" unless -x $visudo;
 
+	# parse sudoers file for "#includedir" directive
+	my $sudodir = parse_sudoers_includedir($sudoers);
+	if ($sudodir) {
+		# sudo will read each file in /etc/sudoers.d, skipping file names that
+		# end in ~ or contain a . character to avoid causing problems with
+		# package manager or editor temporary/backup files
+		$sudoers = "$sudodir/check_raid";
+	}
+
 	warn "Updating file $sudoers\n";
 
 	# NOTE: secure as visudo itself: /etc is root owned
@@ -3803,25 +4043,33 @@ sub sudoers {
 	# setup to have sane perm for new sudoers file
 	umask(0227);
 
-	# insert old sudoers
-	open my $old, '<', $sudoers or die $!;
 	open my $fh, '>', $new or die $!;
-	while (<$old>) {
-		print $fh $_;
-	}
-	close $old or die $!;
 
-	# insert new rules
+	# insert old sudoers
+	if (!$sudodir) {
+		open my $old, '<', $sudoers or die $!;
+		while (<$old>) {
+			print $fh $_;
+		}
+		close $old or die $!;
+	}
+
+	# insert the rules
 	print $fh @rules;
 	close $fh;
 
 	# validate sudoers
 	system($visudo, '-c', '-f', $new) == 0 or unlink($new),exit $? >> 8;
 
-	# use the new file
-	rename($new, $sudoers) or die $!;
-
-	warn "$sudoers file updated.\n";
+	# check if they differ
+	if (filediff($sudoers, $new)) {
+		# use the new file
+		rename($new, $sudoers) or die $!;
+		warn "$sudoers file updated.\n";
+	} else {
+		warn "$sudoers file not changed.\n";
+		unlink($new);
+	}
 }
 
 # Print active plugins
@@ -3857,7 +4105,7 @@ GetOptions(
 	'W' => \$opt_W, 'warnonly' => \$opt_W,
 	'resync=s' => sub { setstate(\$plugin::resync_status, @_); },
 	'check=s' => sub { setstate(\$plugin::check_status, @_); },
-	'noraid=s' => sub { setstate(\$opt_O, @_); },
+	'noraid=s' => sub { setstate(\$noraid_state, @_); },
 	'bbulearn=s' => sub { setstate(\$plugin::bbulearn_status, @_); },
 	'bbu-monitoring' => \$plugin::bbu_monitoring,
 	'p=s' => \$opt_p, 'plugin=s' => \$opt_p,
@@ -3913,7 +4161,11 @@ foreach my $pn (@plugins) {
 		$message .= "$pn:[Plugin error]";
 		next;
 	}
-	$status = $plugin->status if $plugin->status > $status;
+	if ($plugin->message or $noraid_state == $ERRORS{UNKNOWN}) {
+		$status = $plugin->status if $plugin->status > $status;
+	} else {
+		$status = $noraid_state if $noraid_state > $status;
+	}
 	$message .= '; ' if $message;
 	$message .= "$pn:[".$plugin->message."]";
 	$message .= ' | ' . $plugin->perfdata if $plugin->perfdata;
@@ -3931,8 +4183,8 @@ if ($message) {
 		print "UNKNOWN: ";
 	}
 	print "$message\n";
-} elsif ($opt_O) {
-	$status = $ERRORS{OK};
+} elsif ($noraid_state != $ERRORS{UNKNOWN}) {
+	$status = $noraid_state;
 	print "No RAID configuration found\n";
 } else {
 	$status = $ERRORS{UNKNOWN};